+++ app/src/main/java/com/example/trainingapp/view/EditSchedule/EditScheduleActivity.java
@@ -1,0 +1,22 @@
+package com.example.trainingapp.view.EditSchedule;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.trainingapp.R;
+
+/**
+ * Activity for editing the workouts in plans
+ */
+public class EditScheduleActivity extends AppCompatActivity {
+
+    public EditScheduleActivity() {
+        super(R.layout.fragment_edit_schedule);
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+}
