+++ app/src/main/java/com/example/trainingapp/view/EditSchedule/EditWorkout/EditScheduleFragment.java
@@ -1,4 +1,4 @@
-package com.example.trainingapp.view;
+package com.example.trainingapp.view.EditSchedule.EditWorkout;
 
 import android.app.AlertDialog;
 import android.content.DialogInterface;
@@ -7,6 +7,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.TextView;
 
 import androidx.annotation.Nullable;
@@ -15,22 +16,20 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.trainingapp.R;
-import com.example.trainingapp.model.components.Plan;
 import com.example.trainingapp.model.components.Workout;
-import com.example.trainingapp.view.Adapter.EditWorkoutRecyclerViewAdapter;
 import com.example.trainingapp.viewModel.EditWorkoutViewModel;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
- * Fragment for editing workouts during an active session
+ * Fragment for editing the workouts in plans
  */
-public class EditWorkoutFragment extends Fragment {
+public class EditScheduleFragment extends Fragment {
 
     private Workout selectedWorkout;
-    private EditWorkoutRecyclerViewAdapter recyclerViewAdapter;
+    private RecyclerView recyclerView;
+    private EditScheduleRecyclerViewAdapter recyclerViewAdapter;
     private EditWorkoutViewModel editWorkoutViewModel = EditWorkoutViewModel.getInstance();
+
+
     /**
      * onCreateView creates and returns the view hierarchy associated with the fragment.
      *
@@ -46,56 +45,76 @@ public class EditWorkoutFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
 
-        View v = LayoutInflater.from(getContext()).inflate(R.layout.fragment_edit_activeworkout,container,false);
+        View v = LayoutInflater.from(getContext()).inflate(R.layout.fragment_edit_schedule,container,false);
         selectedWorkout = editWorkoutViewModel.getSelectedWorkout();
-        initLabels(v);
+        initTitleText(v);
         initRecyclerView(v);
-        initFinishWorkoutButton(v);
+        initDeleteWorkoutButton(v);
+        initAddExerciseButton(v);
         return v;
     }
 
     /**
-     * initFinishWorkoutButton initiates the 'finish workout'-button
+     * initAddExerciseButton initiates the add exercise button.
      *
      * @param v the current view used in the application
      */
-    private void initFinishWorkoutButton(View v) {
+    private void initAddExerciseButton(View v) {
         Button btnAddExercise = v.findViewById(R.id.btnFinishWorkout);
         btnAddExercise.setVisibility(View.VISIBLE);
-        btnAddExercise.setText("Finish workout");
+        btnAddExercise.setText("Add exercise");
         btnAddExercise.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                if(getActivity().getCurrentFocus() != null) getActivity().getCurrentFocus().clearFocus();
+                editWorkoutViewModel.addExerciseToWorkout(editWorkoutViewModel.getSelectedPlan().getId(), selectedWorkout.getId());
+                recyclerViewAdapter = new EditScheduleRecyclerViewAdapter();
+                recyclerView.setAdapter(recyclerViewAdapter);
+            }
+        });
+    }
+
+    /**
+     * initDeleteWorkoutButton initiates the delete workout button
+     *
+     * @param v the current view used in the application
+     */
+    private void initDeleteWorkoutButton(View v) {
+        Button btnDeleteWorkout = v.findViewById(R.id.btnDeleteWorkout);
+        btnDeleteWorkout.setVisibility(View.VISIBLE);
+        btnDeleteWorkout.setText("Delete workout");
+        btnDeleteWorkout.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
                 initPopupMessageView(v);
             }
         });
     }
 
     /**
-     * initPopupMessageView initiates the popup message view.
+     * initPopupMessageView initiates the popup message view
      *
      * @param v the current view used in the application
      */
     private void initPopupMessageView(View v) {
-        TextView alertTextView = (TextView) v.findViewById(R.id.alertTextViewEndActive);
+        TextView alertTextView = (TextView) v.findViewById(R.id.AlertTextView);
         AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
 
         builder.setCancelable(true);
-        builder.setTitle("Finish workout");
-        builder.setMessage("Are you sure you want to end this workout?");
+        builder.setTitle("Remove workout");
+        builder.setMessage("Are you sure you want to remove this workout?");
 
-        builder.setNegativeButton("KEEP TRAINING", new DialogInterface.OnClickListener() {
+        builder.setNegativeButton("CANCEL", new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialogInterface, int i) {
                 dialogInterface.cancel();
             }
         });
 
-        builder.setPositiveButton("FINISH WORKOUT", new DialogInterface.OnClickListener() {
+        builder.setPositiveButton("REMOVE WORKOUT", new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialogInterface, int i) {
-                editWorkoutViewModel.endActiveWorkout();
+                if (getActivity().getCurrentFocus() != null) getActivity().getCurrentFocus().clearFocus();
+                editWorkoutViewModel.removeWorkoutFromPlan(editWorkoutViewModel.getSelectedPlan().getId(), selectedWorkout.getId());
                 getActivity().finish();
                 alertTextView.setVisibility(View.VISIBLE);
             }
@@ -109,10 +128,10 @@ public class EditWorkoutFragment extends Fragment {
      * @param v the current view used in the application.
      */
     private void initRecyclerView(View v) {
-        RecyclerView recyclerView = (RecyclerView) v.findViewById(R.id.HistoryExerciseRecyclerView);
+        recyclerView = (RecyclerView) v.findViewById(R.id.HistoryExerciseRecyclerView);
         recyclerView.setLayoutManager(new LinearLayoutManager(this.getContext()));
 
-        recyclerViewAdapter = new EditWorkoutRecyclerViewAdapter();
+        recyclerViewAdapter = new EditScheduleRecyclerViewAdapter();
         recyclerView.setAdapter(recyclerViewAdapter);
     }
 
@@ -121,12 +140,21 @@ public class EditWorkoutFragment extends Fragment {
      *
      * @param v the current view used in the application
      */
-    private void initLabels(View v) {
-        TextView tbxWorkoutName = v.findViewById(R.id.tbxWorkoutName);
-        tbxWorkoutName.setText(selectedWorkout.getName());
-        TextView lblReps = v.findViewById(R.id.lblReps);
-        lblReps.setText("Reps");
-        TextView lblWeight = v.findViewById(R.id.lblWeight);
-        lblWeight.setText("Weight");
+    private void initTitleText(View v) {
+        EditText etbxWorkoutName = v.findViewById(R.id.etbxWorkoutName);
+        etbxWorkoutName.setText(selectedWorkout.getName());
+
+        etbxWorkoutName.setOnFocusChangeListener(new View.OnFocusChangeListener() {
+            @Override
+            public void onFocusChange(View view, boolean b) {
+                String txt = etbxWorkoutName.getText().toString();
+                if(txt.equals("")) {
+                    etbxWorkoutName.setText(selectedWorkout.getName() + "");
+                }
+                else {
+                    editWorkoutViewModel.setNewWorkoutName(etbxWorkoutName.getText().toString(), editWorkoutViewModel.getSelectedPlan().getId(), selectedWorkout.getId());
+                }
+            }
+        });
     }
 }
