+++ app/src/main/java/com/example/trainingapp/view/History/HistoryExercise/HistoryFragment.java
@@ -1,4 +1,4 @@
-package com.example.trainingapp.view;
+package com.example.trainingapp.view.History.HistoryExercise;
 
 import java.util.List;
 
@@ -16,7 +16,6 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.example.trainingapp.R;
 import com.example.trainingapp.databinding.FragmentHistoryBinding;
 import com.example.trainingapp.model.activeComponents.ActiveWorkout;
-import com.example.trainingapp.view.Adapter.HistoryRecyclerViewAdapter;
 import com.example.trainingapp.viewModel.HistoryViewModel;
 
 
@@ -26,7 +25,7 @@ import java.util.ArrayList;
  * HistoryFragment acts as the "view" in mvvm. It is responsible for displaying all parts to the
  * fragment.history.xml
  *
- * @author Valdemar Vålvik and Victor Hui
+ * @author Philip Rabia and Patrik Olsson
  */
 
 public class HistoryFragment extends Fragment {
@@ -45,7 +44,7 @@ public class HistoryFragment extends Fragment {
     private FragmentHistoryBinding binding;
 
     private List<ActiveWorkout> testActiveWorkouts = new ArrayList<>();
-    Context context;
+
     HistoryRecyclerViewAdapter recyclerViewAdapter;
 
     /**
