--- app/src/main/java/com/example/trainingapp/view/HistoryExercisesActivity.java
@@ -1,19 +1,0 @@
-package com.example.trainingapp.view;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.example.trainingapp.R;
-
-public class HistoryExercisesActivity extends AppCompatActivity {
-
-    public HistoryExercisesActivity() {
-        super(R.layout.fragment_exercise_history);
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-}
