+++ app/src/main/java/com/example/trainingapp/view/Workout/EditWorkoutActivity.java
@@ -1,4 +1,4 @@
-package com.example.trainingapp.view;
+package com.example.trainingapp.view.Workout;
 
 import android.os.Bundle;
