+++ app/src/main/java/com/example/trainingapp/view/Workout/SelectWorkout/WorkoutRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.example.trainingapp.view.Adapter;
+package com.example.trainingapp.view.Workout.SelectWorkout;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
@@ -8,22 +8,21 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.ListView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.trainingapp.R;
-import com.example.trainingapp.view.EditScheduleActivity;
 import com.example.trainingapp.model.components.Plan;
+import com.example.trainingapp.view.Workout.EditWorkoutActivity;
 import com.example.trainingapp.viewModel.EditWorkoutViewModel;
 
 /**
- * Adapter for the RecyclerView in the ScheduleActivity, provides the correct information for
+ * Adapter for the RecyclerView in the WorkoutActivity, provides the correct information for
  * each list item
  */
-public class ScheduleRecyclerViewAdapter extends RecyclerView.Adapter<ScheduleRecyclerViewAdapter.ListViewHolder>  {
+public class WorkoutRecyclerViewAdapter extends RecyclerView.Adapter<WorkoutRecyclerViewAdapter.ListViewHolder>  {
 
 
     private Plan plan;
@@ -38,12 +37,15 @@ public class ScheduleRecyclerViewAdapter extends RecyclerView.Adapter<ScheduleRe
      * @param context the current state of the application
      * @param activity the current activity
      */
-    public ScheduleRecyclerViewAdapter(Plan plan, Context context, Activity activity) {
+    public WorkoutRecyclerViewAdapter(Plan plan, Context context, Activity activity) {
         this.plan = plan;
         this.context = context;
         this.activity = activity;
     }
 
+    /**
+     * Method for setting a new plan
+     */
     public void setNewPlan(Plan plan) { this.plan = plan; }
 
     @NonNull
@@ -51,7 +53,7 @@ public class ScheduleRecyclerViewAdapter extends RecyclerView.Adapter<ScheduleRe
     public ListViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
 
         return new ListViewHolder(LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.fragment_schedule_list_item, parent, false));
+                .inflate(R.layout.fragment_workout_list_item, parent, false));
     }
 
     /**
@@ -64,14 +66,16 @@ public class ScheduleRecyclerViewAdapter extends RecyclerView.Adapter<ScheduleRe
     public void onBindViewHolder(@NonNull ListViewHolder holder, @SuppressLint("RecyclerView") int position) {
 
         holder.lblWorkoutName.setText(plan.getWorkoutList().get(position).getName());
-        holder.btnEditWorkout.setOnClickListener(new View.OnClickListener() {
+        // set list views
+        holder.btnWorkout.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick (View v) {
+
                 editWorkoutViewModel.setSelectedWorkout(plan.getWorkoutList().get(position));
                 editWorkoutViewModel.setSelectedPlan(plan);
 
                 activity.onBackPressed();
-                Intent openActivity = new Intent(context, EditScheduleActivity.class);
+                Intent openActivity = new Intent(context, EditWorkoutActivity.class);
                 context.startActivity(openActivity);
             }
         });
@@ -85,7 +89,7 @@ public class ScheduleRecyclerViewAdapter extends RecyclerView.Adapter<ScheduleRe
     static class ListViewHolder extends RecyclerView.ViewHolder{
 
         private final TextView lblWorkoutName;
-        private final Button btnEditWorkout;
+        private final Button btnWorkout;
 
         /**
          * Binds elements in layout file to the variables in the ViewHolder
@@ -96,7 +100,8 @@ public class ScheduleRecyclerViewAdapter extends RecyclerView.Adapter<ScheduleRe
             super(itemView);
 
             lblWorkoutName = itemView.findViewById(R.id.lblWorkoutName);
-            btnEditWorkout = itemView.findViewById(R.id.btnWorkout);
+            btnWorkout = itemView.findViewById(R.id.btnWorkout);
+
         }
     }
 }
