+++ app/src/main/res/layout/fragment_history.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#393939"
-    tools:context=".view.HistoryFragment">
+    tools:context=".view.History.HistoryExercise.HistoryFragment">
 
     <TextView
         android:id="@+id/lblCompletedWorkouts"
