+++ app/src/main/res/layout/fragment_workout.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#393939"
-    tools:context=".view.WorkoutFragment">
+    tools:context=".view.Workout.SelectWorkout.WorkoutFragment">
 
     <View
         android:id="@+id/view2"
